refactor: Rename Palette screen file
The filename pattern differed from other screen files by using a period
rather than a dash for the "screen" suffix. This difference resulted in
unexpected triggers of the CHANGELOG CI task due to this native file not
matching the expected native file name pattern matcher:

https://github.com/WordPress/gutenberg/blob/3f3c2cd8f8c0bc40160d295823dbb201a8bdd5f1/.github/workflows/check-components-changelog.yml#L8-L14
