+++ packages/components/src/mobile/color-settings/index.native.js
@@ -13,7 +13,7 @@ import { memo, useEffect, useContext } from '@wordpress/element';
  */
 import PickerScreen from './picker-screen';
 import GradientPickerScreen from './gradient-picker-screen';
-import PaletteScreen from './palette.screen';
+import PaletteScreen from './palette-screen';
 import BottomSheet from '../bottom-sheet';
 import { BottomSheetContext } from '../bottom-sheet/bottom-sheet-context';
